Convert edit-activity tests to Cucumber
